stage4/backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework-stubs==3.16.4
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
